HW3 - Finished up ranks.
Copied in sem related functions from HW2. Modifed sem to be easier to
read and better defined.

rankP (like eval) merely setup up the base state of a rank = 0.
Originally I thought the beginning rank would be the rank of the first
CMD, but that would be a circular definition.

CMD now derives show, and a few test programs were added.