HW3 - Finished semStatTC.
The answer to the question is part b is that by implementing a type
checker, the semantic function no longer has to worry about returning a
maybe data type. This removes all the references to Just and Nothing in
the function and makes it much clearer.

Removed eval function, implemented semantics for new commands, and move
test programs to bottom of section.